Introduced the refinements after review and started Readme file edtion

+++ include/processor.h
@@ -6,7 +6,7 @@ using namespace std::literals::chrono_literals;
 
 class Processor {
  public:
-  float Utilization() const;
+  double Utilization() const;
 
  private:
   const std::chrono::microseconds interval = 200ms;
